requirements.txt
# Core AI/Agent Framework
langchain>=0.1.0
langchain-community>=0.0.20
langchain-groq>=0.1.0
crewai>=0.28.0
groq>=0.4.0

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-multipart>=0.0.6

# Data Validation & Models
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Cardano Integration
pycardano>=0.11.0
blockfrost-python>=0.5.3

# Database & Caching
sqlalchemy>=2.0.0
alembic>=1.13.0
redis>=5.0.0
psycopg2-binary>=2.9.9

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
httpx>=0.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
black>=23.12.0
flake8>=7.0.0
mypy>=1.8.0

# Masumi-specific (if SDK available, otherwise REST client)
# masumi-sdk>=0.1.0  # Uncomment when available
